+++ src/main/java/seedu/inventory/logic/commands/csv/ImportCsvCommand.java
@@ -1,4 +1,4 @@
-package seedu.inventory.logic.commands;
+package seedu.inventory.logic.commands.csv;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.inventory.logic.parser.CliSyntax.PREFIX_FILEPATH;
@@ -8,6 +8,8 @@ import java.util.Objects;
 
 import seedu.inventory.commons.util.FileUtil;
 import seedu.inventory.logic.CommandHistory;
+import seedu.inventory.logic.commands.Command;
+import seedu.inventory.logic.commands.CommandResult;
 import seedu.inventory.logic.commands.exceptions.CommandException;
 import seedu.inventory.model.Model;
