+++ src/main/java/seedu/inventory/logic/parser/InventoryParser.java
@@ -9,16 +9,16 @@ import java.util.regex.Pattern;
 import seedu.inventory.logic.commands.ClearCommand;
 import seedu.inventory.logic.commands.Command;
 import seedu.inventory.logic.commands.ExitCommand;
-import seedu.inventory.logic.commands.ExportCsvCommand;
 import seedu.inventory.logic.commands.HelpCommand;
 import seedu.inventory.logic.commands.HistoryCommand;
-import seedu.inventory.logic.commands.ImportCsvCommand;
 import seedu.inventory.logic.commands.RedoCommand;
 import seedu.inventory.logic.commands.SelectCommand;
 import seedu.inventory.logic.commands.UndoCommand;
 import seedu.inventory.logic.commands.authentication.ChangePasswordCommand;
 import seedu.inventory.logic.commands.authentication.LoginCommand;
 import seedu.inventory.logic.commands.authentication.LogoutCommand;
+import seedu.inventory.logic.commands.csv.ExportCsvCommand;
+import seedu.inventory.logic.commands.csv.ImportCsvCommand;
 import seedu.inventory.logic.commands.item.AddItemCommand;
 import seedu.inventory.logic.commands.item.DeleteItemCommand;
 import seedu.inventory.logic.commands.item.EditItemCommand;
@@ -42,6 +42,8 @@ import seedu.inventory.logic.commands.staff.EditStaffCommand;
 import seedu.inventory.logic.commands.staff.ListStaffCommand;
 import seedu.inventory.logic.parser.authentication.ChangePasswordCommandParser;
 import seedu.inventory.logic.parser.authentication.LoginCommandParser;
+import seedu.inventory.logic.parser.csv.ExportCsvCommandParser;
+import seedu.inventory.logic.parser.csv.ImportCsvCommandParser;
 import seedu.inventory.logic.parser.exceptions.ParseException;
 import seedu.inventory.logic.parser.item.AddCommandParser;
 import seedu.inventory.logic.parser.item.DeleteCommandParser;
