+++ src/main/java/seedu/inventory/storage/csv/CsvAdaptedStaff.java
@@ -1,4 +1,4 @@
-package seedu.inventory.storage;
+package seedu.inventory.storage.csv;
 
 import java.util.ArrayList;
 import java.util.List;
