+++ src/test/java/seedu/inventory/logic/commands/ImportCsvCommandTest.java
@@ -15,6 +15,7 @@ import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
 import seedu.inventory.logic.CommandHistory;
+import seedu.inventory.logic.commands.csv.ImportCsvCommand;
 import seedu.inventory.model.Model;
 import seedu.inventory.model.ModelManager;
 import seedu.inventory.model.SaleList;
