+++ src/test/java/seedu/inventory/logic/parser/InventoryParserTest.java
@@ -24,15 +24,15 @@ import org.junit.rules.ExpectedException;
 
 import seedu.inventory.logic.commands.ClearCommand;
 import seedu.inventory.logic.commands.ExitCommand;
-import seedu.inventory.logic.commands.ExportCsvCommand;
 import seedu.inventory.logic.commands.HelpCommand;
 import seedu.inventory.logic.commands.HistoryCommand;
-import seedu.inventory.logic.commands.ImportCsvCommand;
 import seedu.inventory.logic.commands.RedoCommand;
 import seedu.inventory.logic.commands.SelectCommand;
 import seedu.inventory.logic.commands.UndoCommand;
 import seedu.inventory.logic.commands.authentication.ChangePasswordCommand;
 import seedu.inventory.logic.commands.authentication.LoginCommand;
+import seedu.inventory.logic.commands.csv.ExportCsvCommand;
+import seedu.inventory.logic.commands.csv.ImportCsvCommand;
 import seedu.inventory.logic.commands.item.AddItemCommand;
 import seedu.inventory.logic.commands.item.DeleteItemCommand;
 import seedu.inventory.logic.commands.item.EditItemCommand;
