+++ src/test/java/seedu/inventory/storage/CsvSerializableItemListTest.java
@@ -13,6 +13,8 @@ import org.junit.rules.ExpectedException;
 import seedu.inventory.commons.exceptions.IllegalValueException;
 import seedu.inventory.commons.util.CsvUtil;
 import seedu.inventory.model.ItemList;
+import seedu.inventory.storage.csv.CsvSerializableItemList;
+import seedu.inventory.storage.csv.CsvSerializableSaleList;
 import seedu.inventory.testutil.TypicalItems;
 
 public class CsvSerializableItemListTest {
